refactor: separate the init functions
